+++ CoCowork.DB/StoredProcedures/MiniOfficeOrder/MiniOfficeOrder_Update.sql
@@ -4,7 +4,7 @@
 	@OrderId int,
 	@StartDate datetime,
 	@EndDate datetime,
-	@SubtotalPrice int
+	@SubtotalPrice decimal(10,2)
 AS
 BEGIN
 	update dbo.MiniOfficeOrder
