+++ CoCowork.DB/StoredProcedures/Place/Place_Update.sql
@@ -1,8 +1,8 @@
 ﻿CREATE PROC dbo.Place_Update
 	@Id int,
 	@MiniOfficeId int,
-	@PricePerDay int,
-	@PriceFixedPerDay int
+	@PricePerDay decimal(10,2),
+	@PriceFixedPerDay decimal(10,2)
 AS
 BEGIN
 	update dbo.Place
